ENG-7022: register live pytest options for smoke
